+++ frontend/src/almacen/ShowMisAlmacenes.js
@@ -1,6 +1,8 @@
 import { useEffect, useState } from "react"
 import { useParams, Link } from "react-router-dom"
 import axios from 'axios'
+import almacenDefault from '../imagenes/almacenDefault.jpg'
+
 
 const URIalmacen = 'http://localhost:8000/almacenes/'
 const URIinvitacion = 'http://localhost:8000/invitaciones/'
@@ -23,24 +25,27 @@ const CompShowMisAlmacenes = () => {
     }
 
     const deleteAlmacen = async (id) => {
-        //Borrado en cascada de los objetosAsociados al almacen 
-        const resObj = await axios.get(URIobjeto)
-        let objetos = resObj.data
-        let objetosAsociados = objetos.filter(objeto => objeto.almacenAsociado === id)
-        //Por cada objeto tengo que borrar sus reservas asociadas (luego)
-        objetosAsociados.map(async (objeto) => {
-            await axios.delete(`${URIobjeto}${objeto.id}`)
-        })
+        const confirmarBorrado = window.confirm("¿Estás seguro de que quieres borrar este almacén?");
+        if (confirmarBorrado) {
+            //Borrado en cascada de los objetosAsociados al almacen 
+            const resObj = await axios.get(URIobjeto)
+            let objetos = resObj.data
+            let objetosAsociados = objetos.filter(objeto => objeto.almacenAsociado === id)
+            //Por cada objeto tengo que borrar sus reservas asociadas (luego)
+            objetosAsociados.map(async (objeto) => {
+                await axios.delete(`${URIobjeto}${objeto.id}`)
+            })
 
-        //Borrado en cascada de las invitaciones asociadas al almacen
-        const resInv = await axios.get(URIinvitacion)
-        let invitaciones = resInv.data
-        let invitacionesAsociadas = invitaciones.filter(invitacion => invitacion.almacen === id)
-        invitacionesAsociadas.map(async (invitacion) => {
-            await axios.delete(`${URIinvitacion}${invitacion.id}`)
-        })
+            //Borrado en cascada de las invitaciones asociadas al almacen
+            const resInv = await axios.get(URIinvitacion)
+            let invitaciones = resInv.data
+            let invitacionesAsociadas = invitaciones.filter(invitacion => invitacion.almacen === id)
+            invitacionesAsociadas.map(async (invitacion) => {
+                await axios.delete(`${URIinvitacion}${invitacion.id}`)
+            })
 
-        await axios.delete(`${URIalmacen}${id}`)
+            await axios.delete(`${URIalmacen}${id}`)
+        }
         getAlmacenes()
     }
 
@@ -59,13 +64,13 @@ const CompShowMisAlmacenes = () => {
                     <div className="col" key={almacen.id}>
                         <div className="card text-center mb-4">
                             {almacen.foto == undefined ? 
-                            <img src={process.env.PUBLIC_URL + 'proyecto-interfaces\frontend\src\imagenes\almacenDefault.jpg'} style={{width: '250px', height:'250px', objectFit:'contain'}} className="card-img-top img-fluid"></img>:
-                            <img src={`${almacen.foto}`} style={{width: '250px', height:'250px', objectFit:'contain'}} className="card-img-top img-fluid"></img>
+                            <img src={almacenDefault} style={{width: '250px', height:'250px', objectFit:'contain'}} className="card-img-top img-fluid"></img>:
+                            <img src={almacen.foto} style={{width: '250px', height:'250px', objectFit:'contain'}} className="card-img-top img-fluid"></img>
                             }
                             <div className="card-body">
                                 <h2 className="card-title">{almacen.nombre}</h2>
-                                <Link to={`/${idUser}/${almacen.id}`} className='btn btn-info'>Acceder</Link> 
-                                <button className="ms-2 btn btn-danger" onClick={ ()=>deleteAlmacen(almacen.id)}>Borrar</button>
+                                <Link to={`/${idUser}/${almacen.id}`} className='btn' style={{backgroundColor:'#54A6F0', color: 'black'}}>Acceder</Link> 
+                                <button className="ms-2 btn" style={{backgroundColor:'#EF726B'}} onClick={ ()=>deleteAlmacen(almacen.id)}>Borrar</button>
                             </div>
                         </div>
                     </div>
