+++ frontend/src/miembro/ShowMiembros.js
@@ -84,16 +84,19 @@ const CompShowMiembros = () => {
     }
 
     const expulsarMiembro = async (nombreUsuario) => {
-        const res = await axios.get(URIinvitaciones)
-        let invitados = res.data
-        const res2 = await axios.get(URIusuarios)
-        let usuarios = res2.data
-        let usuario = usuarios.find(usuario => usuario.nombreUsuario == nombreUsuario)
-        let invitacion = invitados.find(invitado => invitado.usuario == usuario.id && invitado.almacen == idAlmacen)
-        console.log(invitados)
-        console.log(invitacion)
-        console.log(invitacion.id)
-        await axios.delete(`${URIinvitaciones}${invitacion.id}`)
+        const confirmarExpulsar = window.confirm("¿Estás seguro de que quieres expulsar a "+nombreUsuario+" de este almacén?");
+        if (confirmarExpulsar) {
+            const res = await axios.get(URIinvitaciones)
+            let invitados = res.data
+            const res2 = await axios.get(URIusuarios)
+            let usuarios = res2.data
+            let usuario = usuarios.find(usuario => usuario.nombreUsuario == nombreUsuario)
+            let invitacion = invitados.find(invitado => invitado.usuario == usuario.id && invitado.almacen == idAlmacen)
+            console.log(invitados)
+            console.log(invitacion)
+            console.log(invitacion.id)
+            await axios.delete(`${URIinvitaciones}${invitacion.id}`)
+        }
         getMiembros()
     }
